Remove agencies lat/lng from `agencies` `POST/PUT` endpoints
