+++ alembic/versions/2025_08_05_1133-0da785aaf39a_remove_agencies_lat_and_lng.py
@@ -5,6 +5,7 @@ Revises: 0896cefcf910
 Create Date: 2025-08-05 11:33:49.811218
 
 """
+
 from typing import Sequence, Union
 
 from alembic import op
@@ -12,17 +13,17 @@ import sqlalchemy as sa
 
 
 # revision identifiers, used by Alembic.
-revision: str = '0da785aaf39a'
-down_revision: Union[str, None] = '0896cefcf910'
+revision: str = "0da785aaf39a"
+down_revision: Union[str, None] = "0896cefcf910"
 branch_labels: Union[str, Sequence[str], None] = None
 depends_on: Union[str, Sequence[str], None] = None
 
 
 def upgrade() -> None:
-    op.drop_column('agencies', 'lat')
-    op.drop_column('agencies', 'lng')
+    op.drop_column("agencies", "lat")
+    op.drop_column("agencies", "lng")
 
 
 def downgrade() -> None:
-    op.add_column('agencies', sa.Column('lat', sa.Float(), nullable=True))
-    op.add_column('agencies', sa.Column('lng', sa.Float(), nullable=True))
+    op.add_column("agencies", sa.Column("lat", sa.Float(), nullable=True))
+    op.add_column("agencies", sa.Column("lng", sa.Float(), nullable=True))
