+++ endpoints/instantiations/agencies_/post/schemas/inner.py
@@ -1,4 +1,4 @@
-from middleware.schema_and_dto.dtos.agencies.post import AgencyInfoPostDTO
+from endpoints.instantiations.agencies_.post.dto import AgencyInfoPostDTO
 from middleware.schema_and_dto.dynamic.pydantic_to_marshmallow.core import (
     pydantic_to_marshmallow,
 )
