+++ middleware/primary_resource_logic/agencies.py
@@ -16,7 +16,7 @@ from middleware.dynamic_request_logic.supporting_classes import (
     PutPostRequestInfo,
 )
 from endpoints.instantiations.agencies_.put.schemas.outer import AgenciesPutSchema
-from middleware.schema_and_dto.dtos.agencies.post import AgenciesPostDTO
+from endpoints.instantiations.agencies_.post.dto import AgenciesPostDTO
 from middleware.schema_and_dto.dtos.agencies.get_many import AgenciesGetManyDTO
 from middleware.schema_and_dto.dtos.common.base import GetByIDBaseDTO
 from middleware.enums import Relations
