+++ middleware/primary_resource_logic/proposals.py
@@ -2,7 +2,7 @@ from db.client.core import DatabaseClient
 from db.enums import ApprovalStatus
 from middleware.security.access_info.primary import AccessInfoPrimary
 from middleware.common_response_formatting import created_id_response
-from middleware.schema_and_dto.dtos.agencies.post import AgenciesPostDTO
+from endpoints.instantiations.agencies_.post.dto import AgenciesPostDTO
 
 
 def propose_agency(
