+++ middleware/schema_and_dto/schemas/agencies/helpers.py
@@ -4,7 +4,7 @@ from marshmallow import fields
 
 from middleware.enums import JurisdictionType
 from endpoints.instantiations.agencies_.put.dto import AgencyInfoPutDTO
-from middleware.schema_and_dto.dtos.agencies.post import AgencyInfoPostDTO
+from endpoints.instantiations.agencies_.post.dto import AgencyInfoPostDTO
 from middleware.schema_and_dto.schemas.agencies.base import AgencyInfoBaseSchema
 from utilities.enums import SourceMappingEnum
