+++ tests/db_client/typeahead/test_agencies.py
@@ -1,6 +1,6 @@
 from db.enums import ApprovalStatus
 from middleware.enums import JurisdictionType, AgencyType
-from middleware.schema_and_dto.dtos.agencies.post import (
+from endpoints.instantiations.agencies_.post.dto import (
     AgenciesPostDTO,
     AgencyInfoPostDTO,
 )
