+++ tests/helpers/helper_functions_complex.py
@@ -17,7 +17,7 @@ from middleware.enums import (
     JurisdictionType,
     AgencyType,
 )
-from middleware.schema_and_dto.dtos.agencies.post import (
+from endpoints.instantiations.agencies_.post.dto import (
     AgencyInfoPostDTO,
     AgenciesPostDTO,
 )
