+++ tests/integration/agencies/test_delete.py
@@ -1,6 +1,8 @@
 from db.db_client_dataclasses import WhereMapping
 from middleware.enums import JurisdictionType, AgencyType
-from middleware.schema_and_dto.schemas.common.common_response_schemas import MessageSchema
+from middleware.schema_and_dto.schemas.common.common_response_schemas import (
+    MessageSchema,
+)
 from tests.helpers.common_test_data import get_test_name
 from tests.helpers.constants import AGENCIES_BASE_ENDPOINT
 from tests.helpers.helper_classes.test_data_creator.flask import TestDataCreatorFlask
